backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
transformers
torch
sentence-transformers
numpy
pdfplumber
python-docx
scikit-learn
nltk
python-dotenv
redis
pytesseract
pillow
psutil
